feat(tarko): extract shared terminal UI components and command highlighting
