+++ multimodal/tarko/agent-web-ui/src/standalone/workspace/components/terminal/index.ts
@@ -1,0 +1,2 @@
+export { TerminalWindow, ScriptTerminalWindow } from './TerminalWindow';
+export { highlightCommand, highlightSimpleCommand } from './commandHighlight';
